Differential Revision: D10416051
Differential Version: 61134800

+++ torch/csrc/cuda/comm.cpp
@@ -149,7 +149,7 @@ std::vector<at::Tensor> scatter(
   }
   at::cuda::CUDAGuard cuda_guard;
   for (size_t chunk = 0; chunk < chunks.size(); ++chunk) {
-    const auto device_index = static_cast<int32_t>(devices[chunk]);
+    const auto device_index = static_cast<int16_t>(devices[chunk]);
     if (streams) {
       AT_CHECK(
           (*streams)[chunk].device() == device_index,
@@ -160,7 +160,7 @@ std::vector<at::Tensor> scatter(
       cuda_guard.set_stream(at::cuda::CUDAStream((*streams)[chunk]));
     }
     chunks[chunk] = chunks[chunk].contiguous().to(
-        {at::DeviceType::CUDA, static_cast<int16_t>(device_index)}, /*non_blocking=*/true);
+        {at::DeviceType::CUDA, device_index}, /*non_blocking=*/true);
   }
   return chunks;
 }
